.opencode/skills/cangjie-dev-harmonyos/scripts/requirements.txt
langchain-chroma>=0.1.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-core>=0.3.0
chromadb>=0.5.0
jieba>=0.42.1
python-dotenv>=1.0.0
